Trim boundary porcelain messages
Exclude git blame entries that exist outside of the requested range (e.g. with --since=<date>). Otherwise, the user with the nearest commit to the boundary earns the LOC count.

Incidentally, this also fixes a crash to stack trace when you request a date-filtered git-fame with --show-email; the auth_stats attributed the boundary lines to a (wrong) author, and that author has no commit messages in the specified range. This can be forced to trigger by requesting a git-fame since the future (e.g. --since=2024) because then one lucky owner gets attributed all of the stat; since all authors have no commits, so there's no emails to perform a lookup against whomever that lucky owner was.